# server/api.yaml
openapi: 3.0.0
servers:
  - url: ./
info:
  title: opensvc collector api
  version: 1.0.3
paths:
...
  /apps/{app_id}:
    post:
      operationId: PostApp
      description: Change an application code properties
      parameters:
        - in: path
          name: app_id
          required: true
          description: App record id or app code
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app:
                  type: string
                description:
                  type: string
                app_domain:
                  type: string
                app_team_ops:
                  type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    get:
      operationId: GetApp
      description: Display app properties
      parameters:
        - in: path
          name: app_id
          required: true
          description: App record id or app code
          schema:
            type: string
        - $ref: '#/components/parameters/inQueryProps'
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteApp
      description: Delete an application code
      parameters:
        - in: path
          name: app_id
          required: true
          description: App record id or app code
          schema:
            type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
...
  /filtersets/{filterset_id}/filters/{f_id}:
    post:
      operationId: PostFiltersetFilter
      description: Attach a filter to a filterset
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
        - in: path
          name: f_id
          required: true
          description: Filter identifier (gen_filters.id or f_label)
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                f_order:
                  type: integer
                f_log_op:
                  type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteFiltersetFilter
      description: Detach a filter from a filterset
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
        - in: path
          name: f_id
          required: true
          description: Filter identifier (gen_filters.id or f_label)
          schema:
            type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
...
  /filtersets/{filterset_id}/filtersets/{child_id}:
    post:
      operationId: PostFiltersetFilterset
      description: Encapsulate a filterset into a filterset
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Parent filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
        - in: path
          name: child_id
          required: true
          description: Child filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                f_order:
                  type: integer
                f_log_op:
                  type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteFiltersetFilterset
      description: Detach an encapsulated filterset from a filterset
      parameters:
        - in: path
          name: filterset_id
          required: true
          description: Parent filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
        - in: path
          name: child_id
          required: true
          description: Child filterset identifier (gen_filtersets.id or fset_name)
          schema:
            type: string
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
  /filtersets_filters:
    post:
      operationId: PostFiltersetsFilters
      description: |
        Attach a filter to a filterset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fset_id
                - f_id
              properties:
                fset_id:
                  type: string
                  description: Filterset record id, or filterset name
                f_id:
                  type: string
                  description: Filter record id, or filter label
                f_log_op:
                  type: string
                  description: Logical operator joining this filter to the previous one
                f_order:
                  type: integer
                  description: Position of the filter in the filterset
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteFiltersetsFilters
      description: |
        Detach a filter from a filterset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fset_id
                - f_id
              properties:
                fset_id:
                  type: string
                  description: Filterset record id, or filterset name
                f_id:
                  type: string
                  description: Filter record id, or filter label
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
  /filtersets_filtersets:
    post:
      operationId: PostFiltersetsFiltersets
      description: |
        Encapsulate a filterset into another filterset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - parent_fset_id
                - child_fset_id
              properties:
                parent_fset_id:
                  type: string
                  description: Parent filterset record id, or name
                child_fset_id:
                  type: string
                  description: Child filterset record id, or name
                f_log_op:
                  type: string
                  description: Logical operator joining this filterset to the previous entry
                f_order:
                  type: integer
                  description: Position of the child filterset in the parent filterset
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
    delete:
      operationId: DeleteFiltersetsFiltersets
      description: |
        Remove a filterset encapsulation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - parent_fset_id
                - child_fset_id
              properties:
                parent_fset_id:
                  type: string
                  description: Parent filterset record id, or name
                child_fset_id:
                  type: string
                  description: Child filterset record id, or name
      tags:
        - collector
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    type: string
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
      security:
        - basicAuth: [ ]
        - bearerAuth: [ ]
...
components:
  responses:
...
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      type: object
      properties:
        text:
          description: |
            A human-readable explanation specific to this occurrence of the
            problem.
          type: string
      required:
        - text
...
  parameters:
...
    inQueryProps:
      in: query
      name: props
      required: false
      description: A list of properties to include in each data dictionnary.
      schema:
        type: string
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT